% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FJSD.R
\name{two_thirds}
\alias{two_thirds}
\title{Calculate the Value at Two-Thirds Position in Sorted Vector}
\usage{
two_thirds(numbers)
}
\arguments{
\item{Numeric}{vector of input values}
}
\value{
Numeric value at the index = length(numbers) * 2/3 after descending sort
}
\description{
Calculate the Value at Two-Thirds Position in Sorted Vector
}
